// src/app/shared/lib/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ListingCondition = {
  NEW: 'NEW',
  LIKE_NEW: 'LIKE_NEW',
  USED: 'USED',
  FOR_PARTS: 'FOR_PARTS'
} as const

export type ListingCondition = (typeof ListingCondition)[keyof typeof ListingCondition]


export const ListingStatus = {
  ACTIVE: 'ACTIVE',
  SOLD: 'SOLD',
  INACTIVE: 'INACTIVE',
  DELETED: 'DELETED'
} as const

export type ListingStatus = (typeof ListingStatus)[keyof typeof ListingStatus]